requirements.txt
bottle
bottle-beaker